lambda/requirements.txt
psycopg2-binary==2.9.9
boto3==1.34.7
requests==2.31.0
